backend/requirements.txt
django
djangorestframework
psycopg2
django-seed
djangorestframework-simplejwt